refactor(fe): api 호출 함수 그룹화 리팩토링
