--- frontend/src/api/comment/deleteComment.ts
@@ -1,22 +1,0 @@
-import { AxiosError } from 'axios';
-
-import { ApiResponse } from '@/types/apiTypes';
-
-import { axiosInstance } from '@/api/axiosConfig';
-import { logError } from '@/utils/auth/errorUtils';
-
-export interface DeleteCommentRequest {
-  commentId: number;
-}
-
-const deleteComment = async (request: DeleteCommentRequest): Promise<ApiResponse<void>> => {
-  try {
-    const response = await axiosInstance.delete(`/api/comments/${request.commentId}`);
-    return response.data;
-  } catch (error: unknown) {
-    logError(error as AxiosError);
-    throw error;
-  }
-};
-
-export default deleteComment;
