+++ frontend/src/api/user/userApi.ts
@@ -6,6 +6,7 @@ import { logError } from '@/utils/auth/errorUtils';
 
 import axiosInstance from '../axiosConfig';
 
+// ======== 비밀번호 검증 ========
 export interface PostVerifyPwdRequest {
   password: string;
 }
@@ -13,6 +14,23 @@ export interface PostVerifyPwdRequest {
 export const postVerifyPwd = async (request: PostVerifyPwdRequest): Promise<ApiResponse<void>> => {
   try {
     const response = await axiosInstance.post('/api/user/password-verification', request);
+    return response.data;
+  } catch (error: unknown) {
+    logError(error as AxiosError);
+    throw error;
+  }
+};
+
+// ======== 계정 탈퇴 ========
+export interface DeleteUserRequest {
+  password: string;
+}
+
+export const deleteUser = async (request: DeleteUserRequest): Promise<ApiResponse<void>> => {
+  try {
+    const response = await axiosInstance.delete('/api/user', {
+      data: request,
+    });
     return response.data;
   } catch (error: unknown) {
     logError(error as AxiosError);
