+++ frontend/src/types/authTypes.ts
@@ -1,0 +1,1 @@
+export type FailReason = 'expired' | 'invalid' | 'already' | 'server';
